ami-core/requirements.txt
ansible==2.7.10
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coloredlogs==10.0
cryptography==2.6.1
docopt==0.6.2
gitdb2==2.0.5
GitPython==2.1.11
humanfriendly==4.18
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
paramiko==2.4.2
psutil==5.6.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
python-dotenv==0.10.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
smmap2==2.0.5
urllib3==1.24.3
